remove unnecessary templates, implement AlterUserinfo and checkCookie, implement notification
Next Step: come up with a method to run course watching
1. crontab
2. backend multi-threading

+++ SCUKiller/views.py
@@ -1,3 +1,4 @@
+from django.http import JsonResponse
 from django.shortcuts import render, HttpResponse, redirect
 
 from django.contrib.auth import login, logout
@@ -66,6 +67,7 @@ def register(request):
                         userProfile.save()
                         user.backend = 'django.contrib.auth.backends.ModelBackend'
                         login(request, user)
+                        CreateNotification(username, "恭喜您！", "您已经成功注册SCUCourseKiller！")
                         return redirect('index')
                 elif captcha != request.session['CheckCode'].lower():
                     errormsg = '验证码错误'
@@ -109,6 +111,7 @@ def logIn(request):
         return render(request, 'login.html', locals())
 
 
+@login_required
 def logOut(request):
     try:
         logout(request)
@@ -125,8 +128,11 @@ def notification(request):
         UserQ = User.objects.get(username=username)
         notifications = noti.objects.filter(host=user)
         notificationsCnt = len(notifications)
-
-        return render(request, 'notification.html', locals())
+        if request.method == 'GET':
+            return render(request, 'notification.html', locals())
+        elif request.method == 'POST':
+            MarkAsRead(request)
+            return redirect('notification')
     else:
         return redirect('login')
 
@@ -137,7 +143,7 @@ def index(request):
         user = request.user
         username = user.username
         UserQ = User.objects.get(username=username)
-        notificationsCnt = len(UserQ.notificationHost.all())
+        notificationsCnt = len(UserQ.notificationHost.filter(isRead=False))
         isNotified = False
         if notificationsCnt != 0:
             isNotified = True
@@ -166,6 +172,16 @@ def inner_index(request):
 def addCourse(request):
     if request.method == 'POST':
         form = AddCourseForm(request.POST)
+        UserQ = User.objects.get(username=request.user.username)
+        userprofile = UserQ.UserProfile
+        try:
+            jwcaccount = userprofile.jwcHost
+        except Exception as e:
+            print(e)
+            jwcaccount = None
+        if jwcaccount is None:
+            notice = '您尚未绑定教务处账号！'
+            return render(request, 'courseManagement.html', locals())
         if form.is_valid():
             UserQ = User.objects.get(username=request.user.username)
             keyword = form.cleaned_data['keyword']
@@ -187,9 +203,16 @@ def addCourse(request):
                         return render(request, 'courseManagement.html', locals())
             host = UserQ.UserProfile
             notice = '课程添加成功！'  # TODO: Need to verify whether it is a vaild course
+            # TODO:如果课程号与课序号都给出则关闭关键词模式
+            if kch != '' and kxh != '':
+                kcm = ''
+                # 都给出时需要获取课程名
+                pass
             course = courses(kch=kch, kxh=kxh, keyword=keyword, host=host, type=ctype)
             course.save()
             Courses = UserQ.UserProfile.coursesHost.all()
+            CreateNotification(username=request.user.username, title="课程添加成功",
+                               content="您已经成功添加课程号为" + str(kch) + "，课序号为" + str(kxh) + "的课程！")
         else:
             notice = '提交的课程信息不合法！'
         return render(request, 'courseManagement.html', locals())
@@ -200,6 +223,8 @@ def jwcAccount(request):
     if request.user.is_authenticated:
         if request.method == 'GET':
             UserQ = User.objects.get(username=request.user.username)
+            # get返回单对象 反向查询需要try
+            # filter返回QuerySet，直接if判断是否为空
             userprofile = UserQ.UserProfile
             try:
                 jwcaccount = userprofile.jwcHost
@@ -223,21 +248,48 @@ def jwcAccount(request):
                     if stuQuery:
                         raise Exception("学号已存在！")
                     try:
-                        cookie = jwcVal.valjwcAccount(stuID, stuPass)
+                        cookie_dict = jwcVal.valjwcAccount(stuID, stuPass)
                         user = User.objects.get(username=request.user.username)
-                        jwc = jwcModel(jwcNumber=stuID, jwcPasswd=stuPass, userprofile=user.UserProfile, jwcCookie=cookie)
+                        jwc = jwcModel(jwcNumber=stuID, jwcPasswd=stuPass, userprofile=user.UserProfile,
+                                       jwcCookie=str(cookie_dict))
                         jwc.save()  # TODO: verify if this is working
+                        CreateNotification(username=request.user.username, title="教务处账号绑定成功",
+                                           content="您已经成功绑定学号为" + str(stuID) + "的教务处账号！")
                         return redirect('jwcAccount')
                     except Exception as e:
                         errormsg = "学号或密码错误！"
                 except Exception as e:
                     errormsg = e
-                return render(request, 'jwcAccount.html', locals())
+                return render(request, 'bindjwcAccount.html', locals())
             else:
                 errormsg = "表单验证未通过！"
                 return render(request, 'bindjwcAccount.html', {'form': form, 'errormsg': errormsg})  # 表单clean验证未通过
     else:
         return redirect('login')
+
+
+@login_required
+def checkCookie(request):
+    if request.method == 'POST':
+        checkNumber = request.POST.get('delNumber')
+        jwcaccount = jwcModel.objects.get(jwcNumber=checkNumber)
+        try:
+            if not jwcaccount:
+                raise Exception("查询的学号不存在！")
+            if request.user.username != jwcaccount.userprofile.user.username:
+                raise Exception("你欲验证的学号不属于你！")
+            cookieStr = jwcaccount.jwcCookie
+            try:
+                valid = jwcVal.valCookie(cookieStr)
+                errormsg = "Cookie有效！"
+            except Exception as e:
+                errormsg = e  # Cookie失效
+        except Exception as e:
+            errormsg = e  # 恶意请求
+        if str(errormsg) == "Cookie已经失效！已经更新为最新的Cookie！":
+            jwcaccount.jwcCookie = str(jwcVal.valjwcAccount(jwcaccount.jwcNumber, jwcaccount.jwcPasswd))
+            jwcaccount.save()
+        return render(request, "jwcAccount.html", locals())
 
 
 @login_required
@@ -251,6 +303,8 @@ def courseManagement(request):
             CourseQ = courses.objects.get(cid=cidDel)
             notice = "课程《" + CourseQ.kcm + "》已被成功删除"
             CourseQ.delete()
+            CreateNotification(username=request.user.username, title="课程删除成功",
+                               content="课程《" + CourseQ.kcm + "》已被成功删除")
             Courses = UserQ.UserProfile.coursesHost.all()
             return render(request, 'courseManagement.html', locals())
 
@@ -275,7 +329,13 @@ def deljwcAccount(request):
                     raise Exception("删除的学号不存在！")
                 if request.user.username != jwcAcc.userprofile.user.username:
                     raise Exception("你欲删除的学号不属于你！")
+                UserQ = User.objects.get(username=request.user.username)
+                remainingCourses = courses.objects.filter(host=UserQ.UserProfile, isSuccess=False)
+                if remainingCourses:
+                    raise Exception("你还有未完成的课程！")
                 jwcAcc.delete()
+                CreateNotification(username=request.user.username, title="教务处账号解绑成功",
+                                   content="您已经成功解绑学号为" + str(delNumber) + "的教务处账号！")
                 errormsg = "解绑成功！"
                 form = AddjwcAccount()
                 return render(request, 'bindjwcAccount.html', locals())
@@ -284,3 +344,47 @@ def deljwcAccount(request):
                 return render(request, 'jwcAccount.html', locals())
     else:
         return redirect('login')
+
+
+@login_required
+def alterUserinfo(request):
+    if request.method == 'GET':
+        username = request.user.username
+        return render(request, 'AlterUserinfo.html', locals())
+    elif request.is_ajax:
+        res = {'status': None, 'msg': None}
+        name = request.POST.get('name')
+        npwd = request.POST.get('npwd')
+        opwd = request.POST.get('opwd')
+        repwd = request.POST.get('re_pwd')
+        user = authenticate(request, username=name, password=opwd)
+        if user:
+            if npwd == repwd:
+                logout(request)
+                user.set_password(npwd)
+                user.save()
+                res['status'] = 200
+                res['msg'] = '密码修改成功'
+
+            else:
+                res['msg'] = '两次密码输入不一致'
+                CreateNotification(username=request.user.username, title="密码修改成功",
+                                   content="您已经成功修改密码！")
+
+        else:
+            res['msg'] = '原密码错误'
+        return JsonResponse(res)
+
+
+def CreateNotification(username, title, content):
+    UserQ = User.objects.get(username=username)
+    notifi = noti(host=UserQ, title=title, content=content)
+    notifi.save()
+
+
+def MarkAsRead(request):
+    nList = noti.objects.filter(host=request.user, isRead=False)
+    for item in nList:
+        if request.POST.get(str(item.id)) == str(item.id):
+            item.isRead = True
+            item.save()
